# Onyx optimizations