--- a/lib/Onyx.js
+++ b/lib/Onyx.js
@@ -373,6 +373,24 @@ function sendDataToConnection(config, val, key) {
 }
 
 /**
+ *  Retrieve all the connections that should be made for a given mapping
+ * @param {object} mapping
+ * @param {string} mapping.key
+ *
+ * @returns {Promise<string[]>}
+ */
+function getKeysToConnectTo(mapping) {
+    if (isCollectionKey(mapping.key)) {
+        // Find all collection items keys matching this mapping
+        return getAllKeys()
+            .then(keys => _.filter(keys, key => isKeyMatch(mapping.key, key)));
+    }
+
+    // Otherwise it's just a single item, wrap it in a list
+    return Promise.resolve([mapping.key]);
+}
+
+/**
  * Subscribes a react component's state directly to a store key
  *
  * @param {object} mapping the mapping information to connect Onyx to the components state
@@ -407,34 +425,27 @@ function connect(mapping) {
             addLastAccessedKey(mapping.key);
         }
     })
-        .then(getAllKeys)
-        .then((keys) => {
-            // Find all the keys matched by the config key
-            const matchingKeys = _.filter(keys, key => isKeyMatch(mapping.key, key));
-
-            // If the key being connected to does not exist, initialize the value with null
+        .then(() => getKeysToConnectTo(mapping))
+        .then((matchingKeys) => {
+            // When there are no matching keys, initialize the value with null
             if (matchingKeys.length === 0) {
                 sendDataToConnection(mapping, null);
-                return;
+                return [];
             }
 
-            // When using a callback subscriber we will trigger the callback
-            // for each key we find. It's up to the subscriber to know whether
-            // to expect a single key or multiple keys in the case of a collection.
-            // React components are an exception since we'll want to send their
-            // initial data as a single object when using collection keys.
+            return multiGet(matchingKeys);
+        })
+        .then((result) => {
+            /* For React components subscribed to collections we want to send the data
+            *  as a single object and trigger setState just once */
             if (mapping.withOnyxInstance && isCollectionKey(mapping.key)) {
-                Promise.all(_.map(matchingKeys, key => get(key)))
-                    .then(values => _.reduce(values, (finalObject, value, i) => ({
-                        ...finalObject,
-                        [matchingKeys[i]]: value,
-                    }), {}))
-                    .then(val => sendDataToConnection(mapping, val));
-            } else {
-                _.each(matchingKeys, (key) => {
-                    get(key).then(val => sendDataToConnection(mapping, val, key));
-                });
+                return sendDataToConnection(mapping, result);
             }
+
+            /* When using a callback subscriber we will trigger the callback
+            *  for each key we find. It's up to the subscriber to know whether
+            *  to expect a single key or multiple keys in the case of a collection. */
+            _.each(result, (val, key) => sendDataToConnection(mapping, val, key));
         });
 
     return connectionID;