--- a/lib/Onyx.js
+++ b/lib/Onyx.js
@@ -4,6 +4,7 @@ import Str from 'expensify-common/lib/str';
 import lodashMerge from 'lodash/merge';
 import {registerLogger, logInfo, logAlert} from './Logger';
 import cache from './OnyxCache';
+import createDeferredTask from './createDeferredTask';
 
 // Keeps track of the last connectionID that was used so we can keep incrementing it
 let lastConnectionID = 0;
@@ -29,11 +30,7 @@ const evictionBlocklist = {};
 let defaultKeyStates = {};
 
 // Connections can be made before `Onyx.init`. They would wait for this promise before resolving
-const deferredInit = {};
-deferredInit.promise = new Promise((res, rej) => {
-    deferredInit.resolve = res;
-    deferredInit.reject = rej;
-});
+const deferredInit = createDeferredTask();
 
 /**
  * Get multiple keys in a single batch