--- a/lib/Onyx.js
+++ b/lib/Onyx.js
@@ -28,6 +28,13 @@ const evictionBlocklist = {};
 // Optional user-provided key value states set when Onyx initializes or clears
 let defaultKeyStates = {};
 
+// Connections can be made before `Onyx.init`. They would wait for this promise before resolving
+const deferredInit = {};
+deferredInit.promise = new Promise((res, rej) => {
+    deferredInit.resolve = res;
+    deferredInit.reject = rej;
+});
+
 /**
  * Get some data from the store
  *
@@ -359,18 +366,20 @@ function connect(mapping) {
         return connectionID;
     }
 
-    // Check to see if this key is flagged as a safe eviction key and add it to the recentlyAccessedKeys list
-    if (mapping.withOnyxInstance && !isCollectionKey(mapping.key) && isSafeEvictionKey(mapping.key)) {
+    // Commit connection only after init passes
+    deferredInit.promise.then(() => {
+        // Check to see if this key is flagged as a safe eviction key and add it to the recentlyAccessedKeys list
+        if (mapping.withOnyxInstance && !isCollectionKey(mapping.key) && isSafeEvictionKey(mapping.key)) {
         // All React components subscribing to a key flagged as a safe eviction
         // key must implement the canEvict property.
-        if (_.isUndefined(mapping.canEvict)) {
+            if (_.isUndefined(mapping.canEvict)) {
             // eslint-disable-next-line max-len
-            throw new Error(`Cannot subscribe to safe eviction key '${mapping.key}' without providing a canEvict value.`);
+                throw new Error(`Cannot subscribe to safe eviction key '${mapping.key}' without providing a canEvict value.`);
+            }
+            addLastAccessedKey(mapping.key);
         }
-        addLastAccessedKey(mapping.key);
-    }
-
-    getAllKeys()
+    })
+        .then(getAllKeys)
         .then((keys) => {
             // Find all the keys matched by the config key
             const matchingKeys = _.filter(keys, key => isKeyMatch(mapping.key, key));
@@ -742,6 +751,9 @@ function init({
         cache.set(key, newValue);
         keyChanged(key, newValue);
     });
+
+    // Give green light to any pending connections
+    deferredInit.resolve();
 }
 
 const Onyx = {