--- a/lib/Onyx.js
+++ b/lib/Onyx.js
@@ -29,6 +29,9 @@ const evictionBlocklist = {};
 // Optional user-provided key value states set when Onyx initializes or clears
 let defaultKeyStates = {};
 
+// Cache cleaning uses this to remove least recently accessed keys
+let MAX_CACHED_KEYS = 150;
+
 // Connections can be made before `Onyx.init`. They would wait for this promise before resolving
 const deferredInit = createDeferredTask();
 
@@ -74,7 +77,11 @@ function multiGet(keys) {
             /* At this point we have everything in cache
             * Instead of searching for the value inside the result, we can retrieved it by key from cache */
             const result = _.chain(keys)
-                .map(key => [key, cache.getValue(key)])
+                .map((key) => {
+                    // eslint-disable-next-line no-use-before-define
+                    addLastAccessedKey(key);
+                    return [key, cache.getValue(key)];
+                })
                 .object()
                 .value();
 
@@ -134,19 +141,6 @@ function isCollectionKey(key) {
 }
 
 /**
- * Find the collection a collection item belongs to
- * or return null if them item is not a part of a collection
- * @param {string} key
- * @returns {string|null}
- */
-function getCollectionKeyForItem(key) {
-    return _.chain(onyxKeys.COLLECTION)
-        .values()
-        .find(name => key.startsWith(name))
-        .value();
-}
-
-/**
  * Checks to see if a given key matches with the
  * configured key of our connected subscriber
  *
@@ -189,7 +183,7 @@ function removeLastAccessedKey(key) {
  */
 function addLastAccessedKey(key) {
     // Only specific keys belong in this list since we cannot remove an entire collection.
-    if (isCollectionKey(key) || !isSafeEvictionKey(key)) {
+    if (isCollectionKey(key)) {
         return;
     }
 
@@ -414,15 +408,13 @@ function connect(mapping) {
 
     // Commit connection only after init passes
     deferredInit.promise.then(() => {
-        // Check to see if this key is flagged as a safe eviction key and add it to the recentlyAccessedKeys list
-        if (mapping.withOnyxInstance && !isCollectionKey(mapping.key) && isSafeEvictionKey(mapping.key)) {
         // All React components subscribing to a key flagged as a safe eviction
         // key must implement the canEvict property.
+        if (mapping.withOnyxInstance && !isCollectionKey(mapping.key) && isSafeEvictionKey(mapping.key)) {
             if (_.isUndefined(mapping.canEvict)) {
-            // eslint-disable-next-line max-len
+                // eslint-disable-next-line max-len
                 throw new Error(`Cannot subscribe to safe eviction key '${mapping.key}' without providing a canEvict value.`);
             }
-            addLastAccessedKey(mapping.key);
         }
     })
         .then(() => getKeysToConnectTo(mapping))
@@ -452,40 +444,26 @@ function connect(mapping) {
 }
 
 /**
- * Remove cache items that are no longer connected through Onyx
- * @param {string} key
+ * Used to periodically clean least recently accessed items from cache
  */
-function cleanCache(key) {
-    const hasRemainingConnections = _.some(callbackToStateMapping, {key});
+function cleanCache() {
+    if (recentlyAccessedKeys.length > MAX_CACHED_KEYS) {
+        // Keep the most recent
+        const mostRecent = recentlyAccessedKeys.splice(recentlyAccessedKeys.length - MAX_CACHED_KEYS, MAX_CACHED_KEYS);
 
-    // When the key is still used in other places don't remove it from cache
-    if (hasRemainingConnections) {
-        return;
+        /* Keep any other items with active connections
+        * Note: after `splice` the array no longer contains the `mostRecent` */
+        const stillConnected = recentlyAccessedKeys.filter((key) => {
+            const hasRemainingConnections = _.some(callbackToStateMapping, mapping => key.startsWith(mapping.key));
+            if (!hasRemainingConnections) {
+                cache.drop(key);
+            }
+
+            return hasRemainingConnections;
+        });
+
+        recentlyAccessedKeys = stillConnected.concat(mostRecent);
     }
-
-    // When this is a collection - also recursively remove any unused individual items
-    if (isCollectionKey(key)) {
-        cache.drop(key);
-
-        getAllKeys().then(cachedKeys => _.chain(cachedKeys)
-            .filter(name => name.startsWith(key))
-            .forEach(cleanCache));
-
-        return;
-    }
-
-    // When this is a collection item - check if the collection is still used
-    const collectionKey = getCollectionKeyForItem(key);
-    if (collectionKey) {
-        // When there's an active subscription for a collection don't remove the item
-        const hasRemainingConnectionsForCollection = _.some(callbackToStateMapping, {key: collectionKey});
-        if (hasRemainingConnectionsForCollection) {
-            return;
-        }
-    }
-
-    // Otherwise remove the value from cache
-    cache.drop(key);
 }
 
 /**
@@ -505,11 +483,10 @@ function disconnect(connectionID, keyToRemoveFromEvictionBlocklist) {
         removeFromEvictionBlockList(keyToRemoveFromEvictionBlocklist, connectionID);
     }
 
-    const key = callbackToStateMapping[connectionID].key;
     delete callbackToStateMapping[connectionID];
 
-    // When the last subscriber disconnects, drop cache as well
-    cleanCache(key);
+    // Check if anything can be removed from cache
+    cleanCache();
 }
 
 /**
@@ -540,7 +517,10 @@ function remove(key) {
  */
 function evictStorageAndRetry(error, ionMethod, ...args) {
     // Find the first key that we can remove that has no subscribers in our blocklist
-    const keyForRemoval = _.find(recentlyAccessedKeys, key => !evictionBlocklist[key]);
+    const keyForRemoval = _.find(
+        recentlyAccessedKeys,
+        key => !evictionBlocklist[key] && isSafeEvictionKey(key),
+    );
 
     if (!keyForRemoval) {
         logAlert('Out of storage. But found no acceptable keys to remove.');
@@ -754,7 +734,9 @@ function mergeCollection(collectionKey, collection) {
 /**
  * Initialize the store with actions and listening for storage events
  *
- * @param {Object} [options]
+ * @param {Object} options
+ * @param {Object} [options.keys]
+ * @param {Object} [options.initialKeyStates]
  * @param {String[]} [options.safeEvictionKeys] This is an array of keys
  * (individual or collection patterns) that when provided to Onyx are flagged
  * as "safe" for removal. Any components subscribing to these keys must also
@@ -762,6 +744,8 @@ function mergeCollection(collectionKey, collection) {
  * @param {function} registerStorageEventListener a callback when a storage event happens.
  * This applies to web platforms where the local storage emits storage events
  * across all open tabs and allows Onyx to stay in sync across all open tabs.
+ * @param {Number} [options.maxCachedKeysCount=150] Sets how many recent keys should we try to keep in cache
+ * Setting this to 0 would only keep active connections in cache
  * @param {Boolean} [options.captureMetrics]
  */
 function init({
@@ -769,12 +753,17 @@ function init({
     initialKeyStates,
     safeEvictionKeys,
     registerStorageEventListener,
+    maxCachedKeysCount,
     captureMetrics = false,
 }) {
     if (captureMetrics) {
         // The code here is only bundled and applied when the captureMetrics is set
         // eslint-disable-next-line no-use-before-define
         applyDecorators();
+    }
+
+    if (_.isNumber(maxCachedKeysCount)) {
+        MAX_CACHED_KEYS = maxCachedKeysCount;
     }
 
     // Let Onyx know about all of our keys