--- a/lib/Onyx.js
+++ b/lib/Onyx.js
@@ -597,11 +597,12 @@ function applyMerge(key, data) {
 function merge(key, val) {
     if (mergeQueue[key]) {
         mergeQueue[key].push(val);
-        return Promise.resolve();
+        return merge.lastPromise;
     }
 
+    // Capture a promise that will resolve when this queue is written
     mergeQueue[key] = [val];
-    return get(key)
+    merge.lastPromise = get(key)
         .then((data) => {
             const modifiedData = applyMerge(key, data);
 
@@ -610,6 +611,8 @@ function merge(key, val) {
             delete mergeQueue[key];
             return set(key, modifiedData);
         });
+
+    return merge.lastPromise;
 }
 
 /**