--- a/lib/Onyx.js
+++ b/lib/Onyx.js
@@ -36,69 +36,37 @@ let MAX_CACHED_KEYS = 150;
 const deferredInit = createDeferredTask();
 
 /**
- * Get multiple keys in a single batch
- *
- * @param {string[]} keys
- * @returns {Promise<Record<string, *>>} - an object of key -> value mapping
- */
-function multiGet(keys) {
-    const [fromCache, hardReads] = _.partition(
-        keys,
-        key => cache.hasCacheForKey(key) || cache.hasPendingTask(`get:${key}`)
-    );
-
-    // Init a list of tasks containing any pending reads from storage
-    const tasks = fromCache
-        .filter(key => cache.hasPendingTask(`get:${key}`))
-        .map(key => Promise.resolve(cache.getTaskPromise(`get:${key}`)));
-
-    // Append a task for any remaining "hard reads" the current list
-    if (hardReads.length > 0) {
-        const readKeysFromStorage = AsyncStorage.multiGet(hardReads)
-            .then(pairs => pairs.forEach(([key, val]) => {
-                const parsed = val && JSON.parse(val);
-                cache.set(key, parsed || null);
-            }));
-
-        tasks.push(readKeysFromStorage);
-
-        /* Capture a resolver for potential multiGet calls that happen while we retrieve keys
-        * example one one call retrieves keys A,B,C another call retrieves B,C,D
-        * we don't need to start a new read for B,C we'll hook them from the first call
-        */
-        hardReads.forEach((key) => {
-            cache.captureTask(`get:${key}`, readKeysFromStorage.then(() => cache.getValue(key)));
-        });
-    }
-
-    // Return a grouped result
-    return Promise.all(tasks)
-        .then(() => {
-            /* At this point we have everything in cache
-            * Instead of searching for the value inside the result, we can retrieved it by key from cache */
-            const result = _.chain(keys)
-                .map((key) => {
-                    // eslint-disable-next-line no-use-before-define
-                    addLastAccessedKey(key);
-                    return [key, cache.getValue(key)];
-                })
-                .object()
-                .value();
-
-            return result;
-        });
-}
-
-/**
  * Get some data from the store
  *
  * @param {string} key
  * @returns {Promise<*>}
  */
 function get(key) {
-    /* We use multiGet as it will batch multiple calls and flush them on the next iteration of
-    * the event loop. AsyncStorage uses `setImmediate` to do this internally */
-    return multiGet([key]).then(result => result[key]);
+    // eslint-disable-next-line no-use-before-define
+    addLastAccessedKey(key);
+
+    // When we already have the value in cache - resolve right away
+    if (cache.hasCacheForKey(key)) {
+        return Promise.resolve(cache.getValue(key));
+    }
+
+    const taskName = `get:${key}`;
+
+    // When a value retrieving task for this key is still running hook to it
+    if (cache.hasPendingTask(taskName)) {
+        return cache.getTaskPromise(taskName);
+    }
+
+    // Otherwise retrieve the value from storage and capture a promise to aid concurrent usages
+    const promise = AsyncStorage.getItem(key)
+        .then((val) => {
+            const parsed = val && JSON.parse(val);
+            cache.set(key, parsed);
+            return parsed;
+        })
+        .catch(err => logInfo(`Unable to get item from persistent storage. Key: ${key} Error: ${err}`));
+
+    return cache.captureTask(taskName, promise);
 }
 
 /**
@@ -367,24 +335,6 @@ function sendDataToConnection(config, val, key) {
 }
 
 /**
- *  Retrieve all the connections that should be made for a given mapping
- * @param {object} mapping
- * @param {string} mapping.key
- *
- * @returns {Promise<string[]>}
- */
-function getKeysToConnectTo(mapping) {
-    if (isCollectionKey(mapping.key)) {
-        // Find all collection items keys matching this mapping
-        return getAllKeys()
-            .then(keys => _.filter(keys, key => isKeyMatch(mapping.key, key)));
-    }
-
-    // Otherwise it's just a single item, wrap it in a list
-    return Promise.resolve([mapping.key]);
-}
-
-/**
  * Subscribes a react component's state directly to a store key
  *
  * @param {object} mapping the mapping information to connect Onyx to the components state
@@ -417,27 +367,34 @@ function connect(mapping) {
             }
         }
     })
-        .then(() => getKeysToConnectTo(mapping))
-        .then((matchingKeys) => {
-            // When there are no matching keys, initialize the value with null
+        .then(getAllKeys)
+        .then((keys) => {
+            // Find all the keys matched by the config key
+            const matchingKeys = _.filter(keys, key => isKeyMatch(mapping.key, key));
+
+            // If the key being connected to does not exist, initialize the value with null
             if (matchingKeys.length === 0) {
                 sendDataToConnection(mapping, null);
-                return [];
+                return;
             }
 
-            return multiGet(matchingKeys);
-        })
-        .then((result) => {
-            /* For React components subscribed to collections we want to send the data
-            *  as a single object and trigger setState just once */
+            // When using a callback subscriber we will trigger the callback
+            // for each key we find. It's up to the subscriber to know whether
+            // to expect a single key or multiple keys in the case of a collection.
+            // React components are an exception since we'll want to send their
+            // initial data as a single object when using collection keys.
             if (mapping.withOnyxInstance && isCollectionKey(mapping.key)) {
-                return sendDataToConnection(mapping, result);
+                Promise.all(_.map(matchingKeys, key => get(key)))
+                    .then(values => _.reduce(values, (finalObject, value, i) => ({
+                        ...finalObject,
+                        [matchingKeys[i]]: value,
+                    }), {}))
+                    .then(val => sendDataToConnection(mapping, val));
+            } else {
+                _.each(matchingKeys, (key) => {
+                    get(key).then(val => sendDataToConnection(mapping, val, key));
+                });
             }
-
-            /* When using a callback subscriber we will trigger the callback
-            *  for each key we find. It's up to the subscriber to know whether
-            *  to expect a single key or multiple keys in the case of a collection. */
-            _.each(result, (val, key) => sendDataToConnection(mapping, val, key));
         });
 
     return connectionID;
@@ -814,7 +771,6 @@ function applyDecorators() {
     // Re-assign with decorated functions
     /* eslint-disable no-func-assign */
     get = decorate.decorateWithMetrics(get, 'Onyx:get');
-    multiGet = decorate.decorateWithMetrics(multiGet, 'Onyx:multiGet');
     set = decorate.decorateWithMetrics(set, 'Onyx:set');
     multiSet = decorate.decorateWithMetrics(multiSet, 'Onyx:multiSet');
     clear = decorate.decorateWithMetrics(clear, 'Onyx:clear');