--- a/lib/Onyx.js
+++ b/lib/Onyx.js
@@ -617,9 +617,12 @@ function merge(key, val) {
 
 /**
  * Merge user provided default key value pairs.
+ *
+ * @returns {Promise<void>}
  */
 function initializeWithDefaultKeyStates() {
-    _.each(defaultKeyStates, (state, key) => merge(key, state));
+    const prepared = prepareKeyValuePairsForStorage(defaultKeyStates);
+    return AsyncStorage.multiMerge(prepared);
 }
 
 /**
@@ -736,17 +739,15 @@ function init({
     evictionAllowList = safeEvictionKeys;
     addAllSafeEvictionKeysToRecentlyAccessedList();
 
-    // Initialize all of our keys with data provided
-    initializeWithDefaultKeyStates();
+    // Initialize all of our keys with data provided then give green light to any pending connections
+    initializeWithDefaultKeyStates()
+        .finally(deferredInit.resolve); // Proceed even if the above task fails
 
     // Update any key whose value changes in storage
     registerStorageEventListener((key, newValue) => {
         cache.set(key, newValue);
         keyChanged(key, newValue);
     });
-
-    // Give green light to any pending connections
-    deferredInit.resolve();
 }
 
 const Onyx = {
@@ -780,6 +781,7 @@ function applyDecorators() {
     merge = decorate.decorateWithMetrics(merge, 'Onyx:merge');
     mergeCollection = decorate.decorateWithMetrics(mergeCollection, 'Onyx:mergeCollection');
     getAllKeys = decorate.decorateWithMetrics(getAllKeys, 'Onyx:getAllKeys');
+    AsyncStorage.multiMerge = decorate.decorateWithMetrics(AsyncStorage.multiMerge, 'Async:multiMerge');
     /* eslint-enable */
 
     // Re-expose decorated methods