<!-- If necessary, assign reviewers that know the area or changes well. Feel free to tag any additional reviewers you see fit. -->
cc @marcaaron @tgolen 

### Details
This PR contains a few optimizations and refactors
1. The first commit has and updated `Onyx.printMetrics` function. You can use it as a base to compare before/after results. It prints a general table on the console which is easy to copy and put in excel. 
2. Fix: some connections happen even before `Onyx.init` is called. This should not be allowed as the `initialKeyStates` are applied during init. Added a deferred tasks that would queue incoming connections and only make them after init, this also gives the rest of the app init some breathing space
3. Perf: Update `set` to check with cache and skip writing to storage. Cache reflects what's already on disc, if we have the exact value in cache it means it didn't change and we don't need to make a write
4. Perf: Updated cache clearing strategy. It's a mix of LRU and active connections. The cache works in the following way: each time a connection is terminated we run a `cleanCache` function. The function removes dated items from cache, if the item is still being used by a component it's kept in cache. By default it will try to keep up to ~150 keys in cache, can be a bit more due to active connections to old items
5. Refactor: `merge` was refactored so that it's promises are resolved when a queue for a key is done. This helps with measuring the function more accurately
6. Perf: Updated `defaultKeyStates` handling. Merging these keys one by one was taking a significant amount of time, even though they are only 4. Instead of writing them to storage, we read data from storage and merge it along with the defaults to the cache. Then we never allow for the `defaultKeys` to be removed from cache 

### Related Issues
<!-- Please replace GH_LINK with the link to the GitHub issue this Pull Request is related to -->
https://github.com/Expensify/Expensify.cash/issues/2667

### Automated Tests
Added and updated tests related to changes in cache handling.
The rest of the changes are covered by the existing tests

### Linked PRs
<!---
Please include links to any update PRs in repos that must change their package.json version.
--->
TBD

### Benchmark results 
There are overall improvements that can be seen in the stats here
https://docs.google.com/spreadsheets/d/1kHRvFL1ITbr4p9TzT_nsmtvCY4KoPwKCXp84FTCsGzk/edit?usp=sharing

##### Android
|                       | Before Updates | After Updates |
|-----------------------|---------------:|--------------:|
| Total Onyx Time       |       55.06sec |      19.67sec |
| Heavy reading ends at |        8.27sec |       5.44sec |

##### iOS
|                       | Before Updates | After Updates |
|-----------------------|---------------:|--------------:|
| Total Onyx Time       |       16.37sec |       8.74sec |
| Heavy reading ends at |        2.96sec |       2.98sec |

- Total: the total time spent by Onyx functions data fetching/writing
- Heavy reading: the point since app launch at which hard reads from disk are over
- The link above contains device information and how the tests were performed