@kidroca 2021-07-14
There's not much to be seen in a visual comparison. I expected to see the LHN much faster, but it seems Onyx is not the main bottleneck ATM 

#### SBS Before/After Android Physical Device
https://user-images.githubusercontent.com/12156624/125672045-fab3eb46-2234-4194-9bef-d092e560cd8d.mp4
- I've trimmed the videos to the point where the JS starts to run, otherwise there is like a ~10sec. splash screen 
- This should be faster with a `prod` build, but I don't have the signing certificate needed to make one

Overall I can feel the app is a bit smoother
There's a noticeable improvement on iOS as well, at one point I've removed the full screen loader, to test a bit and it wasn't even needed - chats rendered instantly

cc @quinthar 
Benchmark info attached in the main description