@kidroca 2021-07-14
Here's the E.cash PR where you can build and try out the changes: https://github.com/Expensify/Expensify.cash/pull/4040 on E.cash